Reworked classification logic and structure. a lot TBD!
